Add loaders, parsers. Correct class Point

+++ TwitUs/src/main/java/org/example/entity/Point.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @NoArgsConstructor
 public class Point {
 
-    private int x;
-    private int y;
+    private Double x;
+    private Double y;
 
 }
